remove ginac
