+++ paddle/cinn/common/CMakeLists.txt
@@ -17,7 +17,6 @@ gather_srcs(
   ir_util.cc
   test_helper.cc
   # cuda_test_helper.cc
-  arithmetic.cc
   cas.cc
   union_find.cc
   python_interpreter_guard.cc
