+++ paddle/cinn/common/cas.cc
@@ -19,7 +19,6 @@
 #include <string>
 #include <utility>
 
-#include "paddle/cinn/common/arithmetic.h"
 #include "paddle/cinn/common/ir_util.h"
 #include "paddle/cinn/ir/ir_mutator.h"
 #include "paddle/cinn/ir/ir_printer.h"
