+++ paddle/cinn/optim/transform_polyfor_to_for.cc
@@ -17,7 +17,6 @@
 #include <cmath>
 #include <vector>
 
-#include "paddle/cinn/common/arithmetic.h"
 #include "paddle/cinn/common/cas.h"
 #include "paddle/cinn/common/ir_util.h"
 #include "paddle/cinn/common/type.h"
